requirements.txt
requests
deep-translator
google-api-python-client 
youtube-transcript-api
tweepy
pandas
python-dotenv
emoji